+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 1.2.25
 
   * Add fetching of email delivery reports -> Osm::Email::Delivery class and subclasses
+  * Add fetching My.SCOUT parent login history - Osm::Myscout::ParentLoginHistory.get_for_section
 
 ## Version 1.2.24